Create LanguageRegistryUtil
